# public/docs/filkom-api-doc.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "API FILKOM-APP",
    "description": "OpenAPI untuk FILKOM APP RESTful API. API ini menggunakan Authentikasi JWT.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:2000/api/v1",
      "description": "Development Server"
    },
    {
      "url": "https://development-filkom-apps-be.vercel.app/api/v1",
      "description": "Deployment Server"
    }
  ],
  "paths": {
...
    "/auth/signin-admin": {
      "post": {
        "tags": ["User Authentication"],
        "summary": "Login Admin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "username": "admin",
                "password": "12345"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Admin Sukses Login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "email": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string"
                            }
                          }
                        },
                        "token": {
                          "type": "string"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "status": "OK",
                  "data": {
                    "user": {
                      "id": "6efa89b3-c086-4038-9ce9-68b99dbcb8bb",
                      "name": "admin",
                      "email": "admin@mail.com",
                      "role": "ADMIN"
                    },
                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7ImlkIjoiNmVmYTg5YjMtYzA4Ni00MDM4LTljZTktNjhiOTlkYmNiOGJiIiwibmFtZSI6ImFkbWluIiwiZW1haWwiOiJhZG1pbkBtYWlsLmNvbSIsInJvbGUiOiJBRE1JTiJ9LCJpYXQiOjE2OTgxOTY5MDZ9.dm6mLjb6FoNulb38DAOSlmOdfFZWRl8qYHPMs1VT3b4"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralError"
                },
                "example": {
                  "status": "FAILED",
                  "data": {
                    "error": "email or password incorrect"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/signin-employee": {
      "post": {
        "tags": ["User Authentication"],
        "summary": "Login Employee",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "username": "kaprodi",
                "password": "12345"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee Sukses Login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "nik": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "role": {
                              "type": "array",
                              "items": {
                                "type": "string"
                              }
                            }
                          }
                        },
                        "token": {
                          "type": "string"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "status": "OK",
                  "data": {
                    "user": {
                      "id": "ab9eb72e-acf5-4497-a7cb-d898ec8cce55",
                      "nik": "kaprodi",
                      "name": "Kaprodi Kaprodi",
                      "role": ["KAPRODI"]
                    },
                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7ImlkIjoiYWI5ZWI3MmUtYWNmNS00NDk3LWE3Y2ItZDg5OGVjOGNjZTU1IiwibmlrIjoia2Fwcm9kaSIsIm5hbWUiOiJLYXByb2RpIEthcHJvZGkiLCJyb2xlIjpbIktBUFJPREkiXX0sImlhdCI6MTY5ODIwMTM0NX0.ZtMAK4H27692KG0SqFZdyVwrpXnnJIs9vF2wNvuyvUA"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralError"
                },
                "example": {
                  "status": "FAILED",
                  "data": {
                    "error": "email or password incorrect"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/signin-student": {
      "post": {
        "tags": ["User Authentication"],
        "summary": "Login Student",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "username": "student",
                "password": "12345"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student Sukses Login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "nim": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "status": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string"
                            }
                          }
                        },
                        "token": {
                          "type": "string"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "status": "OK",
                  "data": {
                    "user": {
                      "id": "91c5d94a-600b-42bd-a470-e35a01f45639",
                      "nim": "student",
                      "name": "John Doe",
                      "status": "ACTIVE",
                      "role": "MAHASISWA"
                    },
                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7ImlkIjoiOTFjNWQ5NGEtNjAwYi00MmJkLWE0NzAtZTM1YTAxZjQ1NjM5IiwibmltIjoic3R1ZGVudCIsIm5hbWUiOiJKb2huIERvZSIsInJvbGUiOlsiTUFIQVNJU1dBIl19LCJpYXQiOjE2OTgyMDE4Nzh9.as4GCkNFq6d5MDoTUUfX9bS4Bv_H7f_jGkOQ-FCjbjI"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralError"
                },
                "example": {
                  "status": "FAILED",
                  "data": {
                    "error": "email or password incorrect"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "filkomAuth": {
        "description": "Authentication for FILKOM APP RESTful API",
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "GeneralError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "error": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  }
}