# src/services/auth/auth.controller.ts
import { zValidator } from "@hono/zod-validator";
import { Hono } from "hono";
import authSchema from "./auth.schema";
import authService from "./auth.service";
import { HTTPException } from "hono/http-exception";

const auth = new Hono().basePath("/auth");

auth.post(
  "/signin-admin",
  zValidator("json", authSchema.login, (res, c) => {
    if (!res.success) {
      throw new HTTPException(400, {
        message: "Please provide valid data",
        cause: res.error,
      });
    }
  }),
  async (c) => {
    const { username, password } = c.req.valid("json");
    const admin = await authService.signInAdmin({ username, password });
    return c.json({ status: "OK", message: "Login Success", data: admin }, 200);
  }
);

auth.post(
  "/signin-employee",
  zValidator("json", authSchema.login, (res, c) => {
    if (!res.success) {
      throw new HTTPException(400, {
        message: "Please provide valid data",
        cause: res.error,
      });
    }
  }),
  async (c) => {
    const { username, password } = c.req.valid("json");
    const employee = await authService.signInEmployee({ username, password });
    return c.json(
      { status: "OK", message: "Login Success", data: employee },
      200
    );
  }
);

auth.post(
  "/signin-student",
  zValidator("json", authSchema.login, (res, c) => {
    if (!res.success) {
      throw new HTTPException(400, {
        message: "Please provide valid data",
        cause: res.error,
      });
    }
  }),
  async (c) => {
    const { username, password } = c.req.valid("json");
    const student = await authService.signInStudent({ username, password });
    return c.json(
      { status: "OK", message: "Login Success", data: student },
      200
    );
  }
);

export default auth;